PostRequestDto - > PostDto 이름 변경 PostService의 update 및 delete 메서드 조건문 추가 - password입력을 통한 객체 확인

+++ page/src/main/java/com/example/page/service/PostService.java
@@ -1,10 +1,9 @@
 package com.example.page.service;
 
-import com.example.page.dtos.PostRequestDtos;
+import com.example.page.dtos.PostDtos;
 import com.example.page.entiity.Post;
 import com.example.page.repository.PostRepository;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -23,7 +22,7 @@ public class PostService {
     }
 
     @Transactional
-    public void createPosting(PostRequestDtos postRequestDtos) {
+    public void createPosting(PostDtos postRequestDtos) {
         Post post = new Post(postRequestDtos);
         postRepository.save(post);
 
@@ -37,19 +36,28 @@ public class PostService {
 
 
     @Transactional
-    public Long modifiedText(Long id, PostRequestDtos postRequestDtos) {
+    public String modifiedText(Long id, PostDtos postRequestDtos, String password) {
         Post post = postRepository.findById(id).orElseThrow(() ->
                 new IllegalArgumentException("등록되지않은 아이디입니다"));
-        post.update(postRequestDtos);
+        if (post.getPassword().equals(password)) {
+            post.update(postRequestDtos);
+            PostDtos postDtos = new PostDtos(post);
+            return postDtos.getName() + "의 글이 수정되었습니다";
+        }
+        return "잘못된 접근입니다";
 //        postRepository.save(post);
-        return post.getId();
     }
 
     @Transactional
-    public void deletePost(Long id, String pw) {
-        Post post = postRepository.findByIdAndPassword(id, pw);
+    public String deletePost(Long id, String pw) {
+        Post post = postRepository.findById(id).orElseThrow(() ->
+                new IllegalArgumentException("해당 id는 저장된 id가 아닙니다"));
+        if (post.getPassword().equals(pw)) {
+            postRepository.delete(post);
+            return id + "값이 삭제되었습니다";
+        }
+        return "올바른 Password가 아닙니다";
 
-        postRepository.delete(post);
 
 
     }
